--- a/benchmark/benchmark/nnlib/upsample.jl
+++ b/benchmark/benchmark/nnlib/upsample.jl
@@ -1,22 +1,23 @@
 ########## upsample ############
 SUITE["nnlib"]["upsample"] = BenchmarkGroup()
 SUITE["nnlib"]["upsample"]["linear"] = BenchmarkGroup()
-for rank in (3, 2, 1,), et in (Float32, Float16,)
-    et_suite = BenchmarkGroup("fw" => BenchmarkGroup(), "bw" => BenchmarkGroup())
-    SUITE["nnlib"]["upsample"]["linear"][string(et)] = et_suite
 
-    inputs_sizes = [
-        (512, (0.5, 2), false), (256, 8, false),
-        (256, 4, true), (128, (1, 2), false), (128, 2, true),
-    ]
-    for (sz, scale, ac) in inputs_sizes
+SIZES = [
+    (3, 256, 8, true),
+    (3, 128, (1, 2), false),
+    (2, 512, (0.5, 2), false),
+    (2, 256, 4, false),
+    (2, 128, 2, true),
+]
+for (rank, sz, scale, ac) in SIZES
+    et_suite = BenchmarkGroup()
+    SUITE["nnlib"]["upsample"]["linear"]["$(rank+2)-N($sz)-scale($scale)"] = et_suite
+    for et in (Float32, Float16,)
+        et_suite[string(et)] = BenchmarkGroup("fw" => BenchmarkGroup(), "bw" => BenchmarkGroup())
         x = ones(et, repeat([sz], rank)..., 1, 1)
-        et_suite["fw"][
-            "$(rank+2)-N($sz)-scale($scale)"
-            ] = @benchmarkable upsample_linear($x, $scale; align_corners = $ac)
-        et_suite["bw"][
-            "$(rank+2)-N($sz)-scale($scale)"
-            ] = @benchmarkable ∇upsample_linear($x;
+        et_suite[string(et)]["fw"] = @benchmarkable upsample_linear($x, $scale;
+                align_corners = $ac)
+        et_suite[string(et)]["bw"] = @benchmarkable ∇upsample_linear($x;
                 size = (typeof($scale) <: Tuple) ?
                     floor.(Integer, $sz .* $scale) :
                     ntuple(_ -> floor(Integer, $sz * $scale), $rank),
@@ -25,7 +26,11 @@ for rank in (3, 2, 1,), et in (Float32, Float16,)
 end
 
 SUITE["nnlib"]["upsample"]["nearest"] = BenchmarkGroup()
-for rank in (3, 2, 1,), N in (512, 256, 128,)
+SIZES = [
+    (3, 256), (2, 512), (2, 256),
+    (1, 512), (1, 256),
+]
+for (rank, N) in SIZES
     et_suite = BenchmarkGroup()
     for et in (Float64, Float32, Float16,)
         x = zeros(Float32, repeat([N], rank)..., 1, 1)