--- a/benchmark/benchmarks.jl
+++ b/benchmark/benchmarks.jl
@@ -13,5 +13,9 @@ register_benchmark(env_name::String, benchmark_file::String) =
 Random.seed!(1234567890)
 const SUITE = BenchmarkGroup()
 
+BenchmarkTools.DEFAULT_PARAMETERS.samples = 20
+BenchmarkTools.DEFAULT_PARAMETERS.seconds = 2.5
+BenchmarkTools.DEFAULT_PARAMETERS.gcsample = true
+
 register_benchmark("FLUXML_BENCHMARK_NNLIB", "benchmark/nnlib.jl")
 register_benchmark("FLUXML_BENCHMARK_FLUX", "benchmark/flux.jl")