--- a/benchmark/runbenchmarks-pr.jl
+++ b/benchmark/runbenchmarks-pr.jl
@@ -15,9 +15,6 @@ baseline_url = parsed_args["baseline"]
 setup_fluxml_env([baseline_url])
 
 using BenchmarkTools
-BenchmarkTools.DEFAULT_PARAMETERS.samples = 20
-BenchmarkTools.DEFAULT_PARAMETERS.seconds = 2.5
-
 using PkgBenchmark
 group_baseline = benchmarkpkg(
     dirname(@__DIR__),
@@ -41,9 +38,6 @@ target_url = parsed_args["target"]
 setup_fluxml_env([target_url])
 
 using BenchmarkTools
-BenchmarkTools.DEFAULT_PARAMETERS.samples = 20
-BenchmarkTools.DEFAULT_PARAMETERS.seconds = 2.5
-
 using PkgBenchmark
 group_target = benchmarkpkg(
     dirname(@__DIR__),