--- a/benchmark/runbenchmarks-pr.jl
+++ b/benchmark/runbenchmarks-pr.jl
@@ -12,11 +12,12 @@ disable_arg = parsed_args["disable"]
 enabled_benchmarks = parse_enabled_benchmarks(enable_arg, disable_arg)
 
 baseline_url = parsed_args["baseline"]
-setup_fluxml_env([baseline_url])
+time_setup_fluxml_env = @elapsed setup_fluxml_env([baseline_url])
+@info "TIME: setup FluxML benchmarking environment (baseline) cost $time_setup_fluxml_env"
 
 using BenchmarkTools
 using PkgBenchmark
-group_baseline = benchmarkpkg(
+time_run_benchmarks = @elapsed begin group_baseline = benchmarkpkg(
     dirname(@__DIR__),
     BenchmarkConfig(
         env = merge(
@@ -25,7 +26,8 @@ group_baseline = benchmarkpkg(
         )
     ),
     resultfile = joinpath(@__DIR__, "result-baseline.json")
-)
+) end
+@info "TIME: run benchmarks (baseline) cost $time_run_benchmarks"
 
 teardown()
 
@@ -35,11 +37,12 @@ Pkg.develop(PackageSpec(path = ENV["PWD"]))
 using FluxMLBenchmarks
 
 target_url = parsed_args["target"]
-setup_fluxml_env([target_url])
+time_setup_fluxml_env = @elapsed setup_fluxml_env([target_url])
+@info "TIME: setup FluxML benchmarking environment (target) cost $time_setup_fluxml_env"
 
 using BenchmarkTools
 using PkgBenchmark
-group_target = benchmarkpkg(
+time_run_benchmarks = @elapsed begin group_target = benchmarkpkg(
     dirname(@__DIR__),
     BenchmarkConfig(
         env = merge(
@@ -48,7 +51,8 @@ group_target = benchmarkpkg(
         )
     ),
     resultfile = joinpath(@__DIR__, "result-target.json"),
-)
+) end
+@info "TIME: run benchmarks (target) cost $time_run_benchmarks"
 
 teardown()