--- a/src/FluxMLBenchmarks.jl
+++ b/src/FluxMLBenchmarks.jl
@@ -9,6 +9,7 @@ export Dependency, get_name, init_dependencies,
     BENCHMARK_PKG_PATH, BENCHMARK_FILES_PATH, FLUXML_AVAILABLE_TOP_LEVEL_BENCHMARKS
 
 include("judge_utils.jl")
-export markdown_report, display_markdown_report
+export markdown_report, display_markdown_report,
+    get_result_files, merge_results
 
 end