### PR Checklist

- [ ] Tests are added
- [ ] Documentation, if applicable

### Description

I've split original benchmarks by GitHub Actions matrix, in order to avoid OOM issues.

<img width="1086" alt="image" src="https://github.com/FluxML/FluxMLBenchmarks.jl/assets/45269589/69d9773e-4c52-4968-aabd-bc8e9bc503f6">

I've validated the feasibility of merging `BenchmarkResults` and exporting the full report.

Since installing dependencies takes amount of time, I add a cache action before running benchmarks.

| Item | Before (s) | After caching (s) |
| :-: | :-: | :-: |
| conv | 354 | 264 |
| upsample | 354 | 222 |
| flux | 322 | 221 |
| ... | ... | ... |

Approximately every matrix item can save 2mins, and Cache job takes 6mins. It can definitely save GitHub Actions quota.